always return uint80 round IDs for FluxAggregator

+++ evm-contracts/src/v0.6/tests/FluxAggregatorTestHelper.sol
@@ -4,7 +4,7 @@ import "../dev/FluxAggregator.sol";
 
 contract FluxAggregatorTestHelper {
 
-  uint32 public requestedRoundId;
+  uint80 public requestedRoundId;
 
   event Here();
